refactor(inventory): Phase 8 reshape Inventory.Tests into Features/
- 10 Api/ test files relocated to Features/<Slice>/ mirroring service slices;
  endpoint files renamed *EndpointTests, namespaces moved to
  Inventory.Tests.Features.<Slice>.
- InternalOutboxEndpointsTests moved to Infrastructure/Outbox/ mirror;
  HealthChecksTests + ObservabilityTests promoted to top-level cross-cutting
  (absorbed the lone Health_WhenCalled test from old InventoryApiTests).
- Domain/ aggregate tests + IntegrationTestBase + factory untouched;
  empty Api/ folder deleted; behaviour unchanged.
