+++ inventory-microservice/Inventory.Tests/HealthChecksTests.cs
@@ -3,13 +3,23 @@ using Microsoft.AspNetCore.TestHost;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
 
-namespace Inventory.Tests.Api;
+namespace Inventory.Tests;
 
 public class HealthChecksTests : IntegrationTestBase
 {
     public HealthChecksTests(InventoryWebApplicationFactory webApplicationFactory)
         : base(webApplicationFactory)
     {
+    }
+
+    [Fact]
+    public async Task Health_WhenCalled_ThenReturnsOk()
+    {
+        // Act
+        var response = await HttpClient.GetAsync("/health");
+
+        // Assert
+        response.EnsureSuccessStatusCode();
     }
 
     [Fact]
